Fix fs.statSync clamping uint64 inodes to INT64_MAX
On filesystems with high 64-bit inode numbers (e.g. NFS mounts with
inodes near 2^63), every file reported the same inode because Stat.zig's
clampedInt64 helper saturated u64 dev/ino/rdev fields to INT64_MAX
before passing them to the C++ binding.

Match Node.js:
- Stats (Number): convert u64 -> f64 via @floatFromInt, losing precision
  above 2^53 but never clamping.
- BigIntStats (BigInt): pass u64 unchanged through a new uint64_t extern
  signature so JSBigInt::createFrom picks the uint64_t overload.

The extern declarations and Bun__createJSStatsObject /
Bun__createJSBigIntStatsObject signatures are updated to match.

+++ src/bun.js/bindings/NodeFSStatBinding.cpp
@@ -596,11 +596,11 @@ JSC::Structure* createJSBigIntStatsObjectStructure(JSC::VM& vm, JSC::JSGlobalObj
     return structure;
 }
 
-extern "C" JSC::EncodedJSValue Bun__createJSStatsObject(Zig::GlobalObject* globalObject, int64_t dev,
-    int64_t ino,
+extern "C" JSC::EncodedJSValue Bun__createJSStatsObject(Zig::GlobalObject* globalObject, double dev,
+    double ino,
     int64_t mode,
     int64_t nlink,
-    int64_t uid, int64_t gid, int64_t rdev, int64_t size, int64_t blksize, int64_t blocks, double atimeMs, double mtimeMs, double ctimeMs, double birthtimeMs)
+    int64_t uid, int64_t gid, double rdev, int64_t size, int64_t blksize, int64_t blocks, double atimeMs, double mtimeMs, double ctimeMs, double birthtimeMs)
 {
     auto& vm = globalObject->vm();
 
@@ -641,13 +641,13 @@ extern "C" JSC::EncodedJSValue Bun__createJSStatsObject(Zig::GlobalObject* globa
 }
 
 extern "C" JSC::EncodedJSValue Bun__createJSBigIntStatsObject(Zig::GlobalObject* globalObject,
-    int64_t dev,
-    int64_t ino,
+    uint64_t dev,
+    uint64_t ino,
     int64_t mode,
     int64_t nlink,
     int64_t uid,
     int64_t gid,
-    int64_t rdev,
+    uint64_t rdev,
     int64_t size,
     int64_t blksize,
     int64_t blocks,
